requirements.txt
python-dotenv == 0.21.0
praw ~= 7.6
google-cloud-storage == 2.16.0